Time in and out of office
Start the program to start your day at the office. Everytime you leave
your desk(because you're a lazy bastard) you just enter a reason and
tell it when you come back. In the end, it substracts all your time that
you didn't work and writes the details to a .txt file.

In brief, is your boss an unforgiving dinosaur? Use this program to
troll them. They'll know when you went to the loo and how many seconds
it took to dump that load. Enjoy. I'm not angry. >:(